Create Python Project - Random Number Guessing Game
Hello! This is my first Python Project about a guessing game where the computer selects a random number within the range of upper and lower bound.
We select the lower and upper bound and the random number lies between those bounds.
User has to guess the number and the no. of guess will be on the basis of the range of the bounds.